tele_requirements.txt
pysocks
logging
telethon